refactor: implement LSM engine core with memtables, compaction, and generic cache support
