+++ src/core/engine/compaction.rs
@@ -1,1 +1,4 @@
-pub struct Compaction;
+#[derive(Debug, Clone, Default)]
+pub struct Compaction {
+    pub max_tables_per_compaction: usize,
+}
