+++ src/core/engine/manifest.rs
@@ -1,1 +1,7 @@
 pub struct Manifest;
+
+impl Manifest {
+    pub fn new() -> Self {
+        Self
+    }
+}
