+++ src/core/mod.rs
@@ -1,2 +1,8 @@
 pub mod engine;
 pub mod log_record;
+pub mod memtable;
+pub mod table;
+pub mod version;
+pub mod iterators;
+pub mod key;
+pub mod cache;
