+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod core;
 pub mod storage;
+pub mod infra;
 
 pub use crate::core::engine::LsmEngine;
