+++ src/storage/sst_iterator.rs
@@ -1,5 +1,6 @@
 use crate::core::engine::LsmEngine;
+use crate::storage::cache::Cache;
 
-pub struct SstIterator {
-    engine: LsmEngine,
+pub struct SstIterator<C: Cache> {
+    engine: LsmEngine<C>,
 }
